Mesh: move PLY reader to own class
